Rewrite familytreemaker.py This took a long time and a lot of work.
The module familytreemaker is now split into three classes - family, household, and person. The classic usage of familytreemaker is now in the family class.
household and person still stand to be rewritten - notably Person is a stripped down version of FamilyTreeMember.

information, marriage, parentage, and familytreemember are all included since I changed those as well - children, partner, and parent now all return the instance of the object rather than the ID of the object. I saw no significant advantage not doing it like this initially.